mongo_benchmark/requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
matplotlib==3.8.2
pandas==2.1.4 